requirements.txt
boto3
pyaudio
websocket-client
asyncio
aiofile
amazon-transcribe
samplerate
numpy
pytictoc